+++ py3dtilers/Texture/texture.py
@@ -78,9 +78,7 @@ class Texture():
             for y in range(0, 3):
                 new_u = (uvs[i][y][0] - offsetX) * ratioX
                 new_v = (uvs[i][y][1] - offsetY) * ratioY
-                # warning : in order to be written correctly, the GLTF writter
-                # expects data to be in float32
-                uvs[i][y] = np.array([new_u, new_v], dtype=np.float32)
+                uvs[i][y] = np.array([new_u, new_v])
 
     @staticmethod
     def get_texture_folder():
